pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bilicache"
version = "0.1.2"
description = "B站视频自动下载缓存工具"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
    { name = "hhxki", email = "hhxki@qq.com" },
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet",
    "Topic :: Multimedia :: Video",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "bilibili-api-python>=5.0.0",
    "aiohttp>=3.7.4,<3.8.0",
    "tomlkit>=0.12.0",
    "segno>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

# ======= 关键：CLI 入口 =======
[project.scripts]
bilicache = "bilicache.cli:main"

# ======= setuptools 配置 =======
[tool.setuptools.packages.find]
where = ["."]
include = ["bilicache*"]
